The Launcher will download its help content from a remote server, e.g.
https://raw.github.com/CVL-dev/cvl-fabric-launcher/master/help/helpfiles.zip

so that the content can be updated without having to release a new
version of the Launcher.

Changes to the files within the helpfiles/ directory will not
immediately affect current Launcher users.  To release updated
help content to current Launchers users, a Launcher administrator
can build a new version of the helpfiles.zip archive and push
it to GitHub.