The wx.STAY_ON_TOP style has been removed from the passphrase dialog, because it was causing problems when trying to open the HTML help window from the passphrase dialog's Help button.
This means that the passphrase dialog is no longer guaranteed
to appear on top of the progress dialog, so the progress
dialog is now hidden while the passphrase dialog is being
displayed, which I think is logical - after all how can
there be progress ticking along while a modal dialog is
being displayed?

Another consequence of removing wx.STAY_ON_TOP from the
passphrase dialog could be that we need to get rid of
shell=True from some calls to subprocess.Popen which could
result in DOS windows popping up (possibly in front of
the passphrase dialog).

KeyDist now takes a parentWindow as an argument, e.g.
the launcherMainFrame.

The passphrase dialog's __init__ and onEnter methods check
whether the parent window is a LauncherMainFrame object, and
if so, they temporarily close the progress dialog, while the
passphrase dialog is being displayed and then reopen it
after the passphrase dialog has closed.

This commit involves changes to both the cvl-ssh-utils submodule
and to the cvl-fabric-launcher repository.