Fixed issue where Launcher didn't grab focus back from TurboVNC after TurboVNC exited on Mac OS X (CVLFAB-590).
Also fixed an issue where the "automatically exit" checkbox
(currently only implemented for the MASSIVE tab, not CVL)
didn't work after the "Stop the Desktop?" dialog was displayed.

Also changed qdel to use the asynchronous option (-a), as was
done in the old launcher.py, instead of using sleep 5, so that
the Launcher exits more quickly when automatically exit is
ticked.  Otherwise if we are making the user wait for as long
as 5 seconds, we would need another progress dialog, so the
user knows why the Launcher is taking so long to exit.