Removed absolute positioning of HelpDialog, e.g. pos=(200,150).  Using CenterOnParent instead.
Changed SimpleOptionDialog's base style from wx.DEFAULT_FRAME_STYLE to wx.DEFAULT_DIALOG_STYLE

Changed 'You don\'t appear to be a member of the project...' message supplied to
ListSelectionDialog for MASSIVE projects, so that it appears on multiple lines,
instead of one really wide line.

If a MASSIVE project is selected from the ListSelectionDialog, updated the MASSIVE project
field in the launcher main frame and in local settings.

MASSIVE Launcher config object and settings file path are now passed into LoginProcess's __init__.

Trying to standardize on "MASSIVE/CVL Launcher" for dialog titles, e.g. instead of 'Submit log?'
Note that in some OS's implementation of wxWidgets, the title might not appear in the title bar
for built-in wx.MessageDialogs - it might appear as bold text within the dialog's main panel.
However, I'm trying to make sure all (or most dialogs) have visible title bars, because otherwise
the user can't drag them out of the way which is annoying, e.g. if they are trying to take
a screenshot of an error behind the borderless dialog to send us.

Parent frame is now passed into ListSelectionDialog, so it can use CenterOnParent(), and
if the parent frame is an object of class LauncherMainFrame, then ListSelectionDialog
can disable the LauncherMainFrame's progress dialog temporarily, so that the ListSelectionDialog
is not competing with the progress dialog for which is the front-most dialog.

ListSelectionDialog now has separate okCallback and cancelCallback.  In enabling the title
bar for dialogs like this one (so the user can move the dialog if desired), I've used the
wx.DEFAULT_DIALOG_STYLE where possible, which also enables an OS-specific close icon within
the title bar.  This event can be caught with EVT_CLOSE, and handled in the same way as
if the user pressed a Cancel button.  And I added a Cancel button to the ListSelectionDialog.

okCallback for ListSelectionDialog (in the case of MASSIVE project selection) now does
extra stuff like updating the MASSIVE project field on the Launcher's main dialog, and
saving the selected MASSIVE project to the user's local settings.

The getProjectsCmd for MASSIVE now only lists MASSIVE projects (with CPU-hour quotas),
instead of listing all Unix groups belonging to the MASSIVE user.  Some groups (e.g.
"namd") may be assigned just to give the user access to a particular application's licence,
and other groups (e.g. "monash") just indicate the user's affiliation, but don't give them
any CPU hour quota to use for a Vis node job.

Enclosed the ListSelectionDialog's widgets within a new listSelectionDialogPanel, which
implements a border (margin) around the edge of the dialog.

Instead of having just one button (OK) on the ListSelectionDialog, which was disabled by
default, there are now 2 buttons (OK and Cancel), which are both enabled, and if the
user clicks OK without making a selection, then they are asked to select an item from
the list.  The actual message they are given is passed into the ListSelectionDialog's
__init__ method, using the noSelectionMessage parameter.