Removed out-of-date comments in WebDAV functions, saying "Only for MASSIVE".
Added a comment that the "displayWebDavAccessInfoInRemoteDialog"
method might not actually display a remote dialog, depending on
the siteConfig.displayWebDavAccessInfoInRemoteDialogCmd.

NOOPCallback in showKillServerDialog has been renamed to
ShutdownKeyDistCallback, because that's what it does now.
The "SHUTDOWN_KEYDIST" event was added because of the
Launcher's new "public_mode" feature, which requires
cleaning up the key.

In the killServer method, the next event is now "SHUTDOWN_KEYDIST",
instead of "SHUTDOWN", again to ensure that keys are cleaned up
properly on shared "Guest" accounts.