Script for automated Mac builds.
It doesn't yet support the pretty DMG background and icon
layout, which result from the manual Mac builds.

There seem to be race conditions or something strange going
on with the Apple Script which tries to make the DMG look
pretty.  The manual method of building the Launcher's Mac
DMG is described on the following page within the CVL's
public wiki:

https://confluence-vre.its.monash.edu.au/display/CVL/MASSIVE-CVL+Launcher+Mac+OS+X+build+instructions