ListSelectionDialog was looking for progressDialog as a property of launcherMainFrame.  This is no longer the case.  The progressDialog is now passed in as an argument, so that it can be temporarily hidden while the modal dialogs is being displayed.
Fixed a bug where the busy cursor sometimes failed to stop after
canceling from the ListSelectionDialog.  This fix required passing
loginprocess as an argument to "showModal", so that "showModal"
can avoid restarting the busy cursor if it detects that the login
process has been canceled or is in the process of being shut down.